Move views, models into packages for extenisbility
