+++ noah_core/models/__init__.py
@@ -1,0 +1,2 @@
+# Import all legacy models
+from .legacy import ItemModel, DonationItemModel, DonationCommitmentModel
