+++ noah_core/urls.py
@@ -1,10 +1,11 @@
 from django.urls import path
 
-from .views import InitView, DonationView, ExportView
+from .views import legacy
 
+# All legacy URL patterns
 urlpatterns = [
-    path('init/', InitView.as_view()),
-    path('donation.register/', DonationView.as_view()),
-
-    path('export.donations/', ExportView.as_view(), name="donations-export")
+    path('init/', legacy.InitView.as_view()),
+    path('donation.register/', legacy.DonationView.as_view()),
+    path('export.donations/', legacy.ExportView.as_view(),
+        name="donations-export")
 ]
